+++ libcxx/test/std/ranges/range.adaptors/range.zip/begin.pass.cpp
@@ -18,7 +18,7 @@
 #include <tuple>
 #include <utility>
 
-#include "types.h"
+#include "../types.h"
 
 template <class T>
 concept HasConstBegin = requires(const T& ct) { ct.begin(); };
