+++ libcxx/test/std/ranges/range.adaptors/range.zip/cpo.pass.cpp
@@ -18,7 +18,7 @@
 #include <type_traits>
 #include <utility>
 
-#include "types.h"
+#include "../types.h"
 
 static_assert(std::is_invocable_v<decltype((std::views::zip))>);
 static_assert(!std::is_invocable_v<decltype((std::views::zip)), int>);
