+++ libcxx/test/std/ranges/range.adaptors/range.zip/ctor.views.pass.cpp
@@ -13,7 +13,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "types.h"
+#include "../types.h"
 
 template <class T>
 void conversion_test(T);
