+++ libcxx/test/std/ranges/range.adaptors/range.zip/end.pass.cpp
@@ -14,7 +14,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "types.h"
+#include "../types.h"
 
 // ID | simple | common | bidi | random | sized | #views |     v.end()    | as_const(v)
 //    |        |        |      | access |       |        |                |   .end()
