+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/arithmetic.pass.cpp
@@ -25,7 +25,7 @@
 #include <concepts>
 #include <functional>
 
-#include "../types.h"
+#include "../../types.h"
 
 template <class T, class U>
 concept canPlusEqual = requires(T& t, U& u) { t += u; };
