+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/compare.pass.cpp
@@ -19,7 +19,7 @@
 #include "test_iterators.h"
 #include "test_range.h"
 
-#include "../types.h"
+#include "../../types.h"
 
 // This is for testing that zip iterator never calls underlying iterator's >, >=, <=, !=.
 // The spec indicates that zip iterator's >= is negating zip iterator's < instead of calling underlying iterator's >=.
