+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/ctor.other.pass.cpp
@@ -17,7 +17,7 @@
 #include <cassert>
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 
 using ConstIterIncompatibleView = BasicView<forward_iterator<int*>, forward_iterator<int*>,
                                             random_access_iterator<const int*>, random_access_iterator<const int*>>;
