+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/decrement.pass.cpp
@@ -16,7 +16,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 
 template <class Iter>
 concept canDecrement = requires(Iter it) { --it; } || requires(Iter it) { it--; };
