+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/deref.pass.cpp
@@ -15,7 +15,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 
 constexpr bool test() {
   std::array a{1, 2, 3, 4};
