+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/increment.pass.cpp
@@ -17,7 +17,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 
 struct InputRange : IntBufferView {
   using IntBufferView::IntBufferView;
