+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/iter_move.pass.cpp
@@ -16,7 +16,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 
 struct ThrowingMove {
   ThrowingMove() = default;
