+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/iter_swap.pass.cpp
@@ -15,7 +15,7 @@
 #include <cassert>
 #include <ranges>
 
-#include "../types.h"
+#include "../../types.h"
 
 struct ThrowingMove {
   ThrowingMove() = default;
