+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/member_types.compile.pass.cpp
@@ -16,7 +16,7 @@
 
 #include "test_iterators.h"
 
-#include "../types.h"
+#include "../../types.h"
 
 template <class T>
 struct ForwardView : std::ranges::view_base {
