+++ libcxx/test/std/ranges/range.adaptors/range.zip/iterator/singular.pass.cpp
@@ -16,7 +16,7 @@
 
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 
 struct ThrowOnIncrementIterator {
   int* it_;
