+++ libcxx/test/std/ranges/range.adaptors/range.zip/range.concept.compile.pass.cpp
@@ -18,7 +18,7 @@
 #include <tuple>
 #include <utility>
 
-#include "types.h"
+#include "../types.h"
 
 void testConceptPair() {
   int buffer1[2] = {1, 2};
