+++ libcxx/test/std/ranges/range.adaptors/range.zip/sentinel/ctor.other.pass.cpp
@@ -13,7 +13,7 @@
 #include <cassert>
 #include <ranges>
 
-#include "../types.h"
+#include "../../types.h"
 
 template <class T>
 struct convertible_sentinel_wrapper {
