+++ libcxx/test/std/ranges/range.adaptors/range.zip/sentinel/eq.pass.cpp
@@ -17,7 +17,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 #include "test_range.h"
 
 using Iterator = random_access_iterator<int*>;
