+++ libcxx/test/std/ranges/range.adaptors/range.zip/sentinel/minus.pass.cpp
@@ -26,7 +26,7 @@
 #include <ranges>
 #include <tuple>
 
-#include "../types.h"
+#include "../../types.h"
 
 template <class Base = int*>
 struct convertible_forward_sized_iterator {
