+++ libcxx/test/std/ranges/range.adaptors/range.zip/size.pass.cpp
@@ -18,7 +18,7 @@
 #include <utility>
 
 #include "test_iterators.h"
-#include "types.h"
+#include "../types.h"
 
 int buffer[] = {1, 2, 3, 4, 5, 6, 7, 8, 9};
 struct View : std::ranges::view_base {
